backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
openai-whisper==20231117
numpy==1.24.3
ffmpeg-python==0.2.0



# fastapi==0.109.0
# uvicorn==0.27.0
# python-multipart==0.0.6
# pydantic==2.5.3
# openai-whisper==20231117
# torch==2.1.2
# torchaudio==2.1.2
# numpy==1.24.3
# setuptools-rust
# ffmpeg-python==0.2.0
